Finaly. JSX-array-iteration launch
"JSX-array-iteration" is a package that will help you with your refactoring path. This is `time saver` in simple words. This package will give you two components. With them you can wrap your React-component or just put it in your code. The Package will iterate over it that times that you specify.